--- src/components/Navigation/Navigation.test.js
@@ -1,9 +1,0 @@
-import React from 'react'
-import renderer from 'react-test-renderer'
-import Header from './Header'
-
-describe('Header component snapshot test', () =>
-  it('renders correctly', () => {
-    const tree = renderer.create(<Header />).toJSON()
-    expect(tree).toMatchSnapshot()
-  }))
